export and unset readygit status!
