+++ Makefile
@@ -6,7 +6,7 @@
 #    By: ismirand <ismirand@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/05/17 16:41:12 by aconceic          #+#    #+#              #
-#    Updated: 2024/08/22 15:44:05 by ismirand         ###   ########.fr        #
+#    Updated: 2024/08/22 16:08:30 by ismirand         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -47,11 +47,12 @@ OBJ = $(addprefix $(OBJ_DIR), $(SRC:.c=.o))
 SRC_DIR = ./source/
 
 SRC = prompt.c frees.c frees2.c lexing/lexing.c lexing/lexing_support.c\
-		lexing/token.c support.c check_input.c check_input2.c signals.c expand/expansion.c\
-		expand/expansion_support.c parsing/parsing.c parsing/tree_debug.c\
-		parsing/tree_support.c parsing/tree_free.c parsing/tree_support2.c\
-		env/env.c init_values.c builtins/echo.c builtins/pwd.c builtins/bt_env.c\
-		builtins/support.c builtins/cd.c builtins/exit.c builtins/export_create.c\
+		lexing/token.c support.c check_input.c check_input2.c signals.c\
+		expand/expansion.c expand/expansion_support.c parsing/parsing.c\
+		parsing/tree_debug.c parsing/tree_support.c parsing/tree_free.c\
+		parsing/tree_support2.c env/env.c init_values.c builtins/echo.c\
+		builtins/pwd.c builtins/bt_env.c builtins/support.c builtins/cd.c\
+		builtins/exit.c builtins/export.c builtins/export_create.c\
 		builtins/unset.c exec/execution.c exec/exec_redir.c exec/exec_cmd.c\
 		exec/exec_pipe.c exec/exec_heredoc.c debug.c main_support.c\
