+++ source/builtins/bt_env.c
@@ -6,7 +6,7 @@
 /*   By: ismirand <ismirand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/29 19:06:41 by aconceic          #+#    #+#             */
-/*   Updated: 2024/08/21 14:45:26 by ismirand         ###   ########.fr       */
+/*   Updated: 2024/08/22 16:10:33 by ismirand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,4 +20,26 @@ void	env(t_env *env_var)
 		ft_putendl_fd(env_var->name, 1);
 		env_var = env_var->next;
 	}
+}
+
+char	*env_join(char *old, char *to_add, t_env *env)
+{
+	char	*temp;
+	char	*temp2;
+	char	*save_name;
+	int		i;
+	
+	i = 0;
+	while (old[i] != '=')
+		i++;
+	save_name = ft_strdup_qt(old, ++i);
+	temp = ft_strtrim(ft_strchr(old, '=') + 1, "\"");
+	free (old);
+	//free (env->name);
+	temp2 = ft_strjoin(save_name, temp);
+	free (save_name);
+	free (temp);
+	env->name = ft_strjoin(temp2, ft_strchr(to_add, '=') + 1);
+	free (temp2);
+	return (env->name);
 }
